feat: Add custom fonts and update navbar menu functionality
The code changes include adding custom fonts to the project and updating the functionality of the navbar menu. The custom fonts "Chillax" and "Satoshi" are added to the project, and the navbar menu's onClick event handler is updated to scroll to the "quem_somos" section. This commit also fixes some style errors in the contact screen and background color.